requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-crontab==0.7.1
django-import-export==3.2.0
et-xmlfile==1.1.0
fontawesomefree==6.4.0
greenlet==2.0.2
idna==3.4
MarkupPy==1.14
numpy==1.24.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.2
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
sqlparse==0.4.4
tablib==3.4.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
xlrd==2.0.1
xlwt==1.3.0
